new testcases which are used for evaluation. also some bug fixes in the code

+++ compiler-tester.py
@@ -8,6 +8,7 @@ Created on Sun Oct 23 00:46:20 2016
 """
 import sys
 import os
+import math
 from tester import compile_code
 from tester import execute_command
 from tester import check_program_output
@@ -17,10 +18,13 @@ import shutil
 def get_assembly_output(a86_dir):
     output_lines = []
     fp = open(a86_dir+"/TEMP.OUT")
-    for line in fp:
-        stripped_line = line.strip("\n")
-        if len(stripped_line) > 0:
-            output_lines.append(stripped_line)
+    try:
+        for line in fp:
+            stripped_line = line.strip("\n")
+            if len(stripped_line) > 0:
+                output_lines.append(stripped_line)
+    except UnicodeDecodeError:
+        return output_lines
     fp.close()
     return output_lines
 
@@ -38,8 +42,17 @@ def get_assembly_output(a86_dir):
 
 def get_line_number_from_error_line(error_line):
     errors = error_line.split(":")
+    # error string does not have good implementation
+    if len(errors) < 2:
+        return 0
     # there should be two strings
-    return int(errors[1].split(" ")[0])
+    line_number = 0
+    try:
+        line_number = int(errors[1].split(" ")[0])
+    except ValueError:
+        return 0
+        
+    return line_number
     
     
 def check_error_lines(program_error_lines, error_output_lines):
@@ -64,6 +77,15 @@ def clean_temp_files(temp_files):
     
 def run_compiler_tests(base_dir, project_dir, a86_dir):
     test_files = get_test_files(base_dir)
+    score = 0
+    test_count = 0
+
+    # if there is output folder remove the folder
+    if os.path.exists(project_dir + '/compiler-output'):
+        shutil.rmtree(project_dir + '/compiler-output')
+    # create output directory
+    os.makedirs(project_dir+'/compiler-output')
+    
     for test_file in test_files:
         full_test_file_path = base_dir + "/testcases/" + test_file
 
@@ -89,10 +111,10 @@ def run_compiler_tests(base_dir, project_dir, a86_dir):
         temp_files.append(full_assembly_file_path)
         temp_files.append(a86_dir+"/TEMP.ASM")
         temp_files.append(a86_dir+"/TEMP.OUT")
-        temp_files.append(a86_dir+"/TEMP.COM")  
+        temp_files.append(a86_dir+"/TEMP.COM")
         
         os.chdir(project_dir)
-        program_output = execute_command("make ARGS=\""+full_test_file_path+"\" run")
+        program_output = execute_command("make run ARGS=\""+full_test_file_path+"\"")
         # convert output to output lines
         program_error_lines = program_output.split("\n")
         # remove the empty line at the end of the program
@@ -106,7 +128,16 @@ def run_compiler_tests(base_dir, project_dir, a86_dir):
                 # continue with the next test file
                 clean_temp_files(temp_files)
                 os.chdir(base_dir)
+                test_count += 1
                 continue
+        
+        # if assembly file is not created just exit        
+        if not os.path.exists(full_assembly_file_path):
+            print(test_file+":FAILED")
+            clean_temp_files(temp_files)
+            os.chdir(base_dir)
+            test_count += 1
+            continue
         
         # copy assembly output to the a86 directory
         shutil.copy(full_assembly_file_path, a86_file_path)
@@ -125,28 +156,37 @@ def run_compiler_tests(base_dir, project_dir, a86_dir):
         
         # process TEMP.OUT
         output_lines = get_assembly_output(a86_dir)
+        # write output in a file
+        output_fp = open('compiler-output/' + test_file, 'w')
+        for line in output_lines:
+            print('[OUTPUT]:' + line)
+            output_fp.write(line + "\n")
+        output_fp.close()
         
         # remove temporary files
         clean_temp_files(temp_files)
-        
+        print('[LENGTHS] output.l:' + str(len(output_lines)) + ' true.l:' + str(len(true_output_lines)))
         if check_program_output(base_dir, true_output_lines, output_lines):
             print(test_file + ":" + "PASSED")
+            score += 1
         else:
             print(test_file + ":" + "FAILED")
         os.chdir(base_dir)
+        test_count += 1
+    print('THE SCORE:' + str(math.ceil((score / float(test_count))*39.0)))
     
 
 def main():
-    if len(sys.argv) != 3:
-        print("tester.py <a86_dir> <project_dir>")
+    if len(sys.argv) != 2:
+        print("tester.py <project_dir>")
         return
     
     # set the base directory of the test script    
     base_dir = os.getcwd()
     # get the directory of assembler dir
-    a86_dir = sys.argv[1]
+    a86_dir = '/home/okan/ta/cmpe230/fall2016/project1/a86/'
     # get directory of the project
-    project_dir = sys.argv[2]
+    project_dir = sys.argv[1]
     
     compile_code(base_dir, project_dir)
     run_compiler_tests(base_dir, project_dir, a86_dir)
